requirements.txt
Flask
Flask-SQLAlchemy
mysqlclient
PyPDF2 
python-docx 
pytesseract 
pdf2image 
Pillow
pymupdf
flask-cors